[feat/#221] :wrench: Refactor: config 경로 변경
