+++ chatty-be/src/main/java/com/chatty/chatty/config/DynamoDBConfig.java
@@ -1,10 +1,9 @@
-package com.chatty.chatty.game.config.dynamodb;
+package com.chatty.chatty.config;
 
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClientBuilder;
-import com.chatty.chatty.config.AWSKey;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
