+++ chatty-be/src/main/java/com/chatty/chatty/config/RestClientConfig.java
@@ -1,4 +1,4 @@
-package com.chatty.chatty.game.config.model;
+package com.chatty.chatty.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
